新增 git ignore 文件，修改 SensorController 的错误拼写，修改 SensorController 初始化时可能造成的空指针异常
