+++ .gitignore
@@ -1,0 +1,5 @@
+*.iml
+.gradle
+*.properties
+.idea
+/build
