+++ app/.gitignore
@@ -1,0 +1,3 @@
+*.iml
+.gradle
+/build
